refactor: remove laana version specification
