+++ src/laana/proto/entity/message.proto
