+++ src/laana/proto/event/buddy.proto
